requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py39h6e9494a_2
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
black=22.1.0=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py39h9ed2024_1003
ca-certificates=2022.3.29=hecd8cb5_0
certifi=2021.10.8=py39hecd8cb5_2
cffi=1.15.0=py39hc55c11b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h6e9494a_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.5.1=py39h9fcab8e_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd3eb1b0_0
dgl=0.7.2=py39_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.2=pyhd8ed1ab_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
giflib=5.2.1=haf1e3a3_0
googledrivedownloader=0.4=pyhd3deb0d_1
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py39hecd8cb5_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.9.1=py39h71a6800_0
ipython=8.0.1=py39h6e9494a_2
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py39h6e9494a_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h9ed2024_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.2=py39h6e9494a_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.2=py39he9d5cce_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=hbcb3906_1
libtiff=4.2.0=h87d7836_0
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=13.0.1=hda6cdc1_0
lz4-c=1.9.3=h23ab428_1
markupsafe=2.0.1=py39h89e85a6_1
matplotlib=3.5.1=py39hecd8cb5_1
matplotlib-base=3.5.1=py39hfb0c5b7_1
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py39h9ed2024_1000
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py39h6e9494a_4
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.3.0=py39hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
networkx=2.6.3=pyhd3eb1b0_0
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
openssl=1.1.1n=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39h4d6be9b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.9.0=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=py39hde71d04_0
pip=21.2.4=py39hecd8cb5_0
platformdirs=2.5.1=pyhd8ed1ab_0
prompt-toolkit=3.0.27=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.0.3=py39_torch_1.10.0_cpu
pygments=2.11.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyrsistent=0.18.0=py39hca72f7f_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.7=h88f2d9e_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.15=pyhd8ed1ab_1
python_abi=3.9=2_cp39
pytorch=1.10.2=py3.9_0
pytorch-cluster=1.5.9=py39_torch_1.10.0_cpu
pytorch-scatter=2.0.9=py39_torch_1.10.0_cpu
pytorch-sparse=0.6.12=py39_torch_1.10.0_cpu
pytorch-spline-conv=1.2.1=py39_torch_1.10.0_cpu
pytz=2021.3=pyhd8ed1ab_0
pyyaml=6.0=py39h89e85a6_3
pyzmq=22.3.0=py39h7fec2f1_1
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.7.3=py39h8c7af03_0
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=h707629a_0
stack_data=0.2.0=pyhd8ed1ab_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.11=h7bc2e8c_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.1=py39h89e85a6_2
tqdm=4.62.3=pyhd8ed1ab_0
traitlets=5.1.1=pyhd8ed1ab_0
typed-ast=1.5.2=py39h89e85a6_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py39hecd8cb5_1
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yacs=0.1.6=py_0
yaml=0.2.5=h0d85af4_2
zeromq=4.3.4=he49afe7_1
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h4dc903c_4
zstd=1.4.9=h322a384_0
